requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.15
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pypdf>=3.17.0
python-docx>=0.8.11
docx2txt>=0.8